separate tests for cmt
