+++ lib/concurrent-merkle-tree/src/lib.rs
@@ -3,5 +3,4 @@ pub mod error;
 pub mod log;
 pub mod merkle_roll;
 pub mod state;
-mod tests;
 pub mod utils;
